add BIOS mode

+++ installer.sh
@@ -2,6 +2,7 @@
 # WARNING: this script will destroy data on the selected disk.
 # This script can be run by executing the following:
 ### curl -sL https://tinyurl.com/24tht5j5 | bash ###
+## if you need to use WiFi use "iwctl" for setup
 
 set -uo pipefail
 trap 's=$?; echo "$0: Error on line "$LINENO": $BASH_COMMAND"; exit $s' ERR
@@ -30,6 +31,13 @@ devicelist=$(lsblk -dplnx size -o name,size | grep -Ev "boot|rpmb|loop" | tac)
 device=$(dialog --stdout --menu "Select installtion disk" 0 0 0 ${devicelist}) || exit 1
 clear
 
+### Check boot mode
+if [ -d /sys/firmware/efi/efivars ]
+  then
+    boot_mode="EFI"
+  else
+    boot_mode="BIOS"
+fi
 
 ### Set up logging ###
 exec 1> >(tee "stdout.log")
@@ -38,35 +46,62 @@ exec 2> >(tee "stderr.log")
 
 timedatectl set-ntp true
 
-
 ### Setup the disk and partitions for GPT/UEFI###
-swap_size=$(free --mebi | awk '/Mem:/ {print $2}')
-swap_end=$(( $swap_size + 129 + 1 ))MiB
+if [ "$boot_mode" == "EFI"]
+  then
+    swap_size=$(free --mebi | awk '/Mem:/ {print $2}')
+    swap_end=$(( $swap_size + 300 + 1 ))MiB
 
-parted --script "${device}" -- mklabel gpt \
-  mkpart ESP fat32 1Mib 129MiB \
-  set 1 boot on \
-  mkpart primary linux-swap 129MiB ${swap_end} \
-  mkpart primary ext4 ${swap_end} 100%
+    parted --script "${device}" -- mklabel gpt \
+      mkpart ESP fat32 1Mib 300MiB \
+      set 1 boot on \
+      mkpart primary linux-swap 129MiB ${swap_end} \
+      mkpart primary ext4 ${swap_end} 100%
 
-# Simple globbing was not enough as on one device I needed to match /dev/mmcblk0p1 
-# but not /dev/mmcblk0boot1 while being able to match /dev/sda1 on other devices.
-part_boot="$(ls ${device}* | grep -E "^${device}p?1$")"
-part_swap="$(ls ${device}* | grep -E "^${device}p?2$")"
-part_root="$(ls ${device}* | grep -E "^${device}p?3$")"
+    # Simple globbing was not enough as on one device I needed to match /dev/mmcblk0p1 
+    # but not /dev/mmcblk0boot1 while being able to match /dev/sda1 on other devices.
+    part_boot="$(ls ${device}* | grep -E "^${device}p?1$")"
+    part_swap="$(ls ${device}* | grep -E "^${device}p?2$")"
+    part_root="$(ls ${device}* | grep -E "^${device}p?3$")"
 
-wipefs "${part_boot}"
-wipefs "${part_swap}"
-wipefs "${part_root}"
+    wipefs "${part_boot}"
+    wipefs "${part_swap}"
+    wipefs "${part_root}"
 
-mkfs.fat -F 32 "${part_boot}"
-mkswap "${part_swap}"
-mkfs.ext4 "${part_root}"
+    mkfs.fat -F 32 "${part_boot}"
+    mkswap "${part_swap}"
+    mkfs.ext4 "${part_root}"
 
-swapon "${part_swap}"
-mount "${part_root}" /mnt
-mount --mkdir "${part_boot}" /mnt/boot
+    swapon "${part_swap}"
+    mount "${part_root}" /mnt
+    mount --mkdir "${part_boot}" /mnt/boot
+fi
 
+### Setup the disk and partitions for MBR/BIOS###
+if [ "$boot_mode" == "BIOS"]
+  then
+    swap_size=$(free --mebi | awk '/Mem:/ {print $2}')
+    swap_end=$(( $swap_size + 1 ))MiB
+
+    parted --script "${device}" -- mklabel msdos \
+      mkpart primary linux-swap 1MiB ${swap_end} \
+      mkpart primary ext4 ${swap_end} 100% \
+      set 2 boot on
+
+    # Simple globbing was not enough as on one device I needed to match /dev/mmcblk0p1 
+    # but not /dev/mmcblk0boot1 while being able to match /dev/sda1 on other devices.
+    part_swap="$(ls ${device}* | grep -E "^${device}p?1$")"
+    part_root="$(ls ${device}* | grep -E "^${device}p?2$")"
+
+    wipefs "${part_swap}"
+    wipefs "${part_root}"
+
+    mkswap "${part_swap}"
+    mkfs.ext4 "${part_root}"
+
+    swapon "${part_swap}"
+    mount "${part_root}" /mnt
+fi
 
 ### Install and configure the basic system ###
 # cat >>/etc/pacman.conf <<EOF
@@ -77,7 +112,7 @@ mount --mkdir "${part_boot}" /mnt/boot
 
 # pacstrap /mnt mdaffin-desktop
 pacstrap /mnt base base-devel linux linux-firmware intel-ucode nano sudo networkmanager git
-genfstab -t PARTUUID /mnt >> /mnt/etc/fstab
+genfstab -U /mnt >> /mnt/etc/fstab
 
 
 echo "${hostname}" > /mnt/etc/hostname
@@ -104,8 +139,9 @@ arch-chroot /mnt locale-gen
 echo "LANG=en_US.UTF-8" >> /mnt/etc/locale.conf
 echo "KEYMAP=de-latin1" >> /mnt/etc/vconsole.conf
 
-
-# installing the boot loader
+### installing the boot loader for GPT/UEFI ###
+if [ "$boot_mode" == "EFI"]
+then
 arch-chroot /mnt bootctl --path=/boot install
 
 cat <<EOF > /mnt/boot/loader/loader.conf
@@ -119,6 +155,14 @@ initrd   /intel-ucode.img
 initrd   /initramfs-linux.img
 options  root=PARTUUID=$(blkid -s PARTUUID -o value "$part_root") rw
 EOF
+fi
+
+if [ "$boot_mode" == "BIOS"]
+  then
+    pacstrap /mnt grub
+    arch-chroot /mnt grub-install --target=i386-pc "${device}"
+    arch-chroot /mnt grub-mkconfig -o /boot/grub/grub.cfg
+fi
 
 ### adding the user ###
 arch-chroot /mnt useradd -mG wheel "$user"
